requirements.txt
discord.py
requests
colorama
console
requests-toolbelt
faker
